+++ js/main.js
@@ -4,7 +4,6 @@ import { openForm } from './form.js';
 import './hashtags-pristine.js';
 import './effects.js';
 import { loadData } from './fetch.js';
-import { showAlert } from './util.js';
 import './filters.js';
 import './add-photos.js';
 
@@ -17,7 +16,17 @@ const onSuccess = (data) => {
 };
 
 const onError = () => {
-  showAlert();
+  const messageAlert = document.createElement('div');
+  messageAlert.style.position = 'absolute';
+  messageAlert.style.left = 0;
+  messageAlert.style.top = 0;
+  messageAlert.style.right = 0;
+  messageAlert.style.fontSize = '20px';
+  messageAlert.style.backgroundColor = '#e1375f';
+  messageAlert.style.padding = '15px';
+  messageAlert.style.textAlign = 'center';
+  messageAlert.textContent = 'Ошибка загрузки фотографий';
+  document.body.append(messageAlert);
 };
 
 loadData(onSuccess, onError);
