+++ js/util.js
@@ -34,18 +34,5 @@ const closeOnEscKeyDown = (evt, cb) => {
   }
 };
 
-const showAlert = () => {
-  const messageAlert = document.createElement('div');
-  messageAlert.style.position = 'absolute';
-  messageAlert.style.left = 0;
-  messageAlert.style.top = 0;
-  messageAlert.style.right = 0;
-  messageAlert.style.fontSize = '20px';
-  messageAlert.style.backgroundColor = '#e1375f';
-  messageAlert.style.padding = '15px';
-  messageAlert.style.textAlign = 'center';
-  messageAlert.textContent = 'Ошибка загрузки фотографий';
-  document.body.append(messageAlert);
-};
 
-export { closeOnEscKeyDown, debounce, shuffleArray, isEscapeKey, showAlert };
+export { closeOnEscKeyDown, debounce, shuffleArray, isEscapeKey };
